워치 앱 '켜자마자 꺼짐' 크래시 수정 — 밀리초 Int→Int64 (arm64_32)
증상: 꼼모닝 워치앱 열면 자꾸 꺼짐.
원인: WatchPush.syncToken()이 푸시 토큰을 RTDB에 기록할 때 `t: Int(ms)`로 넣었는데,
  애플워치(arm64_32)는 Int가 32비트라 밀리초(≈1.75조)가 Int.max(21억)를 초과 → 변환 트랩(크래시).
  이 함수가 토큰 수신 직후(런치 직후) 호출돼서 앱이 켜자마자 꺼진 것. (FireREST엔 이미 Int64 가드 있었는데 WatchPush만 빠져있었음.)
수정: Int64로. JSONSerialization도 Int64 OK. 워치 코드 전수 스캔 — 다른 Int 오버플로·as!·강제언랩 없음.

+++ ios/App/KkomWatch/WatchPush.swift
@@ -45,7 +45,9 @@ final class WatchPush: NSObject, UNUserNotificationCenterDelegate {
     func syncToken() {
         guard let hex = lastTokenHex, let r = role,
               let url = URL(string: "\(RTDB)/watchTokens/\(keyForName(r)).json") else { return }
-        let body: [String: Any] = ["token": hex, "platform": "watchos", "t": Int(Date().timeIntervalSince1970 * 1000)]
+        // ⚠️ 애플워치(arm64_32)는 Int가 32비트 → 밀리초(≈1.75조)를 Int로 넣으면 Int.max 초과로 '켜자마자 크래시'.
+        //   토큰 등록은 런치 직후 실행돼서 앱이 열자마자 꺼졌다. Int64 필수(JSONSerialization도 Int64 OK).
+        let body: [String: Any] = ["token": hex, "platform": "watchos", "t": Int64(Date().timeIntervalSince1970 * 1000)]
         var req = URLRequest(url: url); req.httpMethod = "PUT"
         req.setValue("application/json", forHTTPHeaderField: "Content-Type")
         req.httpBody = try? JSONSerialization.data(withJSONObject: body)
